:art:rename: 컴포넌트 공통 및 페이지별 폴더 생성하여 분리
